+++ packages/uikit/src/components/jettons/TronFeeBanner.tsx
@@ -316,7 +316,7 @@ const FeeTable = () => {
                             {t('tron_fee_table_free_transfer_subtitle_used_next_on_date', {
                                 date: formatDate(trc20FreeTransfers.rechargeDate, {
                                     day: 'numeric',
-                                    month: 'short'
+                                    month: 'long'
                                 })
                             })}
                         </TableSecondLineText>
